fix DB
