+++ app/Models/Category.php
@@ -19,10 +19,5 @@ class Category extends Model
     {
         return $this->hasMany(Product::class);
     }
-
-    public function sale()
-    {
-        return $this->hasOne(Sale::class);
-    }
 }
