+++ app/Models/Color.php
@@ -5,21 +5,12 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Sale extends Model
+class Color extends Model
 {
     use SoftDeletes;
 
     protected $fillable = [
-        'code',
-        'product_id',
         'name',
-        'description',
-        'formality',
-        'percent',
-        'amount',
-        'type',
-        'start_date',
-        'end_date',
     ];
 
     public $timestamps = true;
