--- app/Models/ProductDetail.php
@@ -1,26 +1,0 @@
-<?php
-
-namespace App\Models;
-
-use Illuminate\Database\Eloquent\Model;
-
-class ProductDetail extends Model
-{
-    protected $fillable = [
-        'product_id',
-        'quantity',
-        'color',
-    ];
-
-    public $timestamps = true;
-
-    public function product()
-    {
-        return $this->belongsTo(Product::class);
-    }
-
-    public function orders()
-    {
-        return $this->belongsTomany(Order::class);
-    }
-}
