+++ app/Models/ProductInformation.php
@@ -5,21 +5,14 @@ namespace App\Models;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Sale extends Model
+class ProductDetail extends Model
 {
     use SoftDeletes;
 
     protected $fillable = [
-        'code',
-        'product_id',
         'name',
+        'brand',
         'description',
-        'formality',
-        'percent',
-        'amount',
-        'type',
-        'start_date',
-        'end_date',
     ];
 
     public $timestamps = true;
