+++ app/Models/Supplier.php
@@ -17,7 +17,7 @@ class Supplier extends Model
 
     public $timestamps = true;
 
-    public function product()
+    public function products()
     {
         return $this->belongsToMany(Product::class);
     }
