+++ database/seeds/UserSeeder.php
@@ -16,6 +16,7 @@ class UserSeeder extends Seeder
                 'name' => 'User',
                 'email' => 'user@gmail.com',
                 'password' => '$2y$10$fSz9qggDZ2LYzXFzYHLTmOGEUzUmmO5bmBuHaBf7fPX3wppSareEG',
+                'avatar' =>'storage/avatar_default.png',
                 'role_id' => config('setting.role.user'),
                 'created_at' => new DateTime,
                 'updated_at' => new DateTime,
@@ -24,6 +25,7 @@ class UserSeeder extends Seeder
                 'name' => 'Admin Product',
                 'email' => 'admin.product@gmail.com',
                 'password' => '$2y$10$fSz9qggDZ2LYzXFzYHLTmOGEUzUmmO5bmBuHaBf7fPX3wppSareEG',
+                'avatar' =>'storage/avatar_default.png',
                 'role_id' => config('setting.role.admin_product'),
                 'created_at' => new DateTime,
                 'updated_at' => new DateTime,
@@ -32,6 +34,7 @@ class UserSeeder extends Seeder
                 'name' => 'Admin Order',
                 'email' => 'admin.order@gmail.com',
                 'password' => '$2y$10$fSz9qggDZ2LYzXFzYHLTmOGEUzUmmO5bmBuHaBf7fPX3wppSareEG',
+                'avatar' =>'storage/avatar_default.png',
                 'role_id' => config('setting.role.admin_order'),
                 'created_at' => new DateTime,
                 'updated_at' => new DateTime,
@@ -40,6 +43,7 @@ class UserSeeder extends Seeder
                 'name' => 'Management',
                 'email' => 'management@gmail.com',
                 'password' => '$2y$10$fSz9qggDZ2LYzXFzYHLTmOGEUzUmmO5bmBuHaBf7fPX3wppSareEG',
+                'avatar' =>'storage/avatar_default.png',
                 'role_id' => config('setting.role.management'),
                 'created_at' => new DateTime,
                 'updated_at' => new DateTime,
